+++ src/krotov/__init__.py
@@ -14,6 +14,23 @@ estimators for second-order updates.
 
 __version__ = '0.2.0+dev'
 
+__arxiv__ = '1902.11284'
+
+__citation__ = (
+    "M. H. Goerz et al., Krotov: A Python implementation of Krotov's method for quantum optimal control, arXiv:%s (2019)"
+    % __arxiv__
+)
+
+__bibtex__ = r'''
+@article{arxiv1902.11284,
+    author = {Michael H. Goerz and Daniel Basilewitsch and Fernando Gago-Encinas and Matthias G. Krauss and Karl P. Horn and Daniel M. Reich and Christiane P. Koch},
+    title = {Krotov: A {Python} implementation of {Krotov's} method for quantum optimal control},
+    year = {2019},
+    journal = {arXiv:1902.11284},
+}
+'''.strip()
+
+
 # expose submodules for easy import
 from . import shapes
 from . import structural_conversions